nxtcrawler/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Automat==20.2.0
awscli==1.25.71
backcall==0.2.0
backports.zoneinfo;python_version<"3.9"
banal==1.0.6
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
boto==2.49.0
boto3==1.24.70
botocore==1.27.70
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
chompjs==1.1.8
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.6.0
dataset==1.5.2
dateparser==1.1.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docutils==0.16
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
frozenlist==1.3.1
greenlet==1.1.3
h11==0.14.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==5.0.0
importlib-resources==5.10.0
incremental==21.3.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kombu==5.2.4
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.4.12
outcome==1.2.0
packaging==21.3
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
premailer==3.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
Protego==0.2.1
psutil==5.9.2
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
queuelib==1.6.2
redis==4.5.3
regex==2022.3.2
requests==2.28.1
requests-file==1.5.1
retrying==1.3.3
rfc3987==1.3.8
rsa==4.7.2
s3transfer==0.6.0
schematics==2.1.1
scrapinghub==2.4.0
Scrapy==2.6.2
scrapy-selenium==0.0.7
selenium==4.6.0
Send2Trash==1.8.0
sentry-sdk==1.9.8
service-identity==21.1.0
six==1.16.0
slackclient==2.9.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spidermon==1.17.0
SQLAlchemy==1.4.41
stack-data==0.5.1
strict-rfc3339==0.7
terminado==0.16.0
text-unidecode==1.3
tinycss2==1.1.1
tldextract==3.4.0
tornado==6.2
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.8.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
w3lib==2.0.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
widgetsnbextension==4.0.3
wsproto==1.2.0
yarl==1.8.1
zipp==3.9.0
zope.interface==5.4.0
